- move API calls to api module
